+++ src/tools/types/MatrixInterface.java
@@ -19,7 +19,7 @@ public interface MatrixInterface {
     // Jika matriks memiliki 4 kolom, maka getColEff() akan mengembalikan nilai 4.
 
     // Mendapatkan elemen matriks pada baris i dan kolom j
-    float getElmt(int i, int j);
+    double getElmt(int i, int j);
     // Input: Baris (i) dan Kolom (j)
     // Output: Nilai elemen pada baris i dan kolom j
     // Contoh Case:
@@ -57,7 +57,7 @@ public interface MatrixInterface {
     // Jika matriks memiliki 4 kolom, maka getFirstIdxCol() akan mengembalikan nilai 0.
 
     // Mendapatkan elemen diagonal pada baris i
-    float getElmtDiagonal(int i);
+    double getElmtDiagonal(int i);
     // Input: Baris (i)
     // Output: Nilai elemen diagonal pada baris i
     // Contoh Case:
@@ -93,7 +93,7 @@ public interface MatrixInterface {
     // Jika matriks adalah matriks identitas, maka isIdentity() akan mengembalikan true.
 
     // Menetapkan nilai elemen matriks pada baris i dan kolom j
-    void setElmt(int i, int j, float value);
+    void setElmt(int i, int j, double value);
     // Input: Baris (i), Kolom (j), dan Nilai (value)
     // Output: Tidak ada
     // Contoh Case:
@@ -109,7 +109,7 @@ public interface MatrixInterface {
     // true.
 
     // Menyalin elemen-elemen matriks ke dalam array 2D baru
-    float[][] copyElmtMatrix();
+    double[][] copyElmtMatrix();
     // Input: Tidak ada
     // Output: Array 2D baru yang berisi salinan elemen-elemen matriks
     // Contoh Case:
@@ -141,7 +141,7 @@ public interface MatrixInterface {
 
     Matrix multiplyMatrix(Matrix m1, Matrix m2);
 
-    void multiplyByConst(float k);
+    void multiplyByConst(double k);
 
     void negation();
 
@@ -149,19 +149,19 @@ public interface MatrixInterface {
 
     Matrix createIdentityMatrix(int rows, int cols);
 
-    void OBEPlusRow(int idxRowOrigin, int idxRowTarget, float factor);
+    void OBEPlusRow(int idxRowOrigin, int idxRowTarget, double factor);
 
     void OBESwapRow(int idxRowOrigin, int idxRowTarget);
 
-    void OBEDivisionFactor(int i, float factor);
+    void OBEDivisionFactor(int i, double factor);
 
     void gaussElimination();
 
     void gJordanElimination();
 
-    float determinantCofactor();
+    double determinantCofactor();
 
-    float determinantUpperTriangle();
+    double determinantUpperTriangle();
 
     Matrix cofactor();
